requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.1
Django==3.2.5
django-bootstrap4==2.0.1
django-filter==2.3.0
djangorestframework==3.11.0
Markdown==3.2.2
Pillow==8.3.2
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
python-dotenv==0.19.0
